Add dynamic_cast asserts to CPU Loops.
On CUDA, we currently support casting loops dynamically (i.e. when the argument or return types of the lamba don't match the dtypes of the TensorIterator).
On CPU, before this change we would essentially reinterpret_cast, now we internal assert.  We could add dynamic_casting support in the future on CPU.

ghstack-source-id: 8e6b69c33f2ff5b8a106a69f457d676ed2d003f3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/38817