Add "batching rule" for torch.sum(tensor, dims)
On the way to moving vmap prototype into tree, this is the
implementation of the sum batching rule and some utility functions.
After implementing this and another batching rule (probably for
at::mul), we'll be ready to for the initial vmap Python
API.

Re: testing: There's a bit of a chicken-and-egg problem here; I'd like to test
that the batching rules work for vmap, but the easiest way to do that is using
the (non-existent) vmap python API. This PR includes some minimal testing for
this batching rule.

Future PRs (for more context):
- [YOU ARE HERE] Add torch.sum(tensor, dims) batching rule + some helper functions
- Add torch.mul batching rule + some more helper functions
- Add vmap python API
- Add "slow" for-loop fallbacks for out-of-place functions via
dispatcher fallback mechanism.
- Add batching rules for "view" functions
- Add "slow" for-loop fallbacks for in-place functions
- And more

Test Plan:
- `./build/bin/vmap_test`

ghstack-source-id: c8d66c5da7cfdae59f3a82a6220de5ab7d11f24b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/38723