Fix and reenable threaded QNNPACK linear
Summary:
Previously, this was causing divide-by-zero only in the multithreaded
empty-batch case, while calculating tiling parameters for the threads.
In my opinion, the bug here is using a value that is allowed to be zero
(batch size) for an argument that should not be zero (tile size), so I
fixed the bug by bailing out right before the call to
pthreadpool_compute_4d_tiled.

Test Plan:
TestQuantizedOps.test_empty_batch

ghstack-source-id: c8c0e2aaac676849230483ef57c99df956c35356
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40587