Fix batch size zero for QNNPACK linear_dynamic
Summary:
Two bugs were preventing this from working.  One was a divide by zero
when multithreading was enabled, fixed similarly to the fix for static
quantized linear in the previous commit.  The other was computation of
min and max to determine qparams.  FBGEMM uses [0,0] for [min,max] of
empty input, do the same.

Test Plan:
Added a unit test.

ghstack-source-id: d93e650abe9f6b4143d3c8e79cdc504413d9c59e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40588