Enable `out_dims` for vmap frontend API
`out_dims` specifies where in the output tensors the vmapped dimension
should appear. We implement this by simply creating a view with the
batch dimension moved to the desired position.

`out_dims` must either:
- be int (use the same value for all outputs)
- be Tuple[int] (so the user specifies one out_dim per output).
(See the vmap docstring for what we advertise out_dims to do).

I also renamed `TestVmap` to `TestVmapAPI` to make it clearer that we
are testing the API here and not specific operators (which will go into
their own test class).

Test Plan:
- `pytest test/test_vmap.py -v`

ghstack-source-id: 735b20228f5dd8c1302f739d819bdaf1804b32c9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40576